Offer MongoURL guidance in options
I struggled to get my MongoDB configuration working (I kept being told it wasn't configured when it clearly was!)

I saw elsewhere that IndieKit expects MongoDB URIs to have a path element (to define the database name), I added `/indiekit` to my URI and things started working.

There may be ways to set this as a default if it's unspecified, but I'm not familiar enough with the codebase to offer a fix yet. Feel free to replace this PR with a code-based one instead!